+++ tests/drivers/clock_control/clock_control_api/src/test_clock_control.c
@@ -9,52 +9,12 @@
 LOG_MODULE_REGISTER(test);
 
 #if DT_HAS_COMPAT_STATUS_OKAY(nordic_nrf_clock)
-#include <zephyr/drivers/clock_control/nrf_clock_control.h>
+#include "nrf_device_subsys.h"
+#elif DT_HAS_COMPAT_STATUS_OKAY(espressif_esp32_rtc)
+#include "esp32_device_subsys.h"
+#else
+#error "Unsupported board"
 #endif
-
-struct device_subsys_data {
-	clock_control_subsys_t subsys;
-	uint32_t startup_us;
-};
-
-struct device_data {
-	const struct device *dev;
-	const struct device_subsys_data *subsys_data;
-	uint32_t subsys_cnt;
-};
-
-#if DT_HAS_COMPAT_STATUS_OKAY(nordic_nrf_clock)
-static const struct device_subsys_data subsys_data[] = {
-	{
-		.subsys = CLOCK_CONTROL_NRF_SUBSYS_HF,
-		.startup_us =
-			IS_ENABLED(CONFIG_SOC_SERIES_NRF91X) ?
-				3000 : 500
-	},
-#ifndef CONFIG_SOC_NRF52832
-	/* On nrf52832 LF clock cannot be stopped because it leads
-	 * to RTC COUNTER register reset and that is unexpected by
-	 * system clock which is disrupted and may hang in the test.
-	 */
-	{
-		.subsys = CLOCK_CONTROL_NRF_SUBSYS_LF,
-		.startup_us = (CLOCK_CONTROL_NRF_K32SRC ==
-			NRF_CLOCK_LFCLK_RC) ? 1000 : 500000
-	}
-#endif /* !CONFIG_SOC_NRF52832 */
-};
-#endif /* DT_HAS_COMPAT_STATUS_OKAY(nordic_nrf_clock) */
-
-static const struct device_data devices[] = {
-#if DT_HAS_COMPAT_STATUS_OKAY(nordic_nrf_clock)
-	{
-		.dev = DEVICE_DT_GET_ONE(nordic_nrf_clock),
-		.subsys_data =  subsys_data,
-		.subsys_cnt = ARRAY_SIZE(subsys_data)
-	}
-#endif
-};
-
 
 typedef void (*test_func_t)(const struct device *dev,
 			    clock_control_subsys_t subsys,
@@ -194,7 +154,9 @@ static bool async_capable(const struct device *dev, clock_control_subsys_t subsy
 	int err;
 
 	err = clock_control_async_on(dev, subsys, async_capable_callback, NULL);
-	if (err < 0) {
+	if (err == -ENOSYS) {
+		ztest_test_skip();
+	} else if (err < 0) {
 		printk("failed %d", err);
 		return false;
 	}
