+++ tests/drivers/clock_control/clock_control_api/testcase.yaml
@@ -1,8 +1,16 @@
+common:
+  tags:
+    - drivers
+    - clock_control
 tests:
+  drivers.clock.clock_control_api_default:
+    platform_allow:
+      - esp32_devkitc_wroom/esp32/procpu
+      - esp32_devkitc_wrover/esp32/procpu
+      - esp32c3_devkitm
+      - esp32s2_saola
+      - esp32s3_devkitm/esp32s3/procpu
   drivers.clock.clock_control_nrf5:
-    tags:
-      - drivers
-      - clock_control
     platform_allow:
       - nrf51dk/nrf51822
       - nrf52dk/nrf52832
@@ -12,9 +20,6 @@ tests:
     integration_platforms:
       - nrf51dk/nrf51822
   drivers.clock.clock_control_nrf5_lfclk_rc:
-    tags:
-      - drivers
-      - clock_control
     platform_allow:
       - nrf51dk/nrf51822
       - nrf52dk/nrf52832
